put route handlers & authentication

+++ backend/db/migrations/20221022040955-create-spot.js
@@ -27,10 +27,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       lat: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.DOUBLE
       },
       lng: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.DOUBLE
       },
       name: {
         type: Sequelize.STRING
